# src/main/resources/oas/openapi.yaml
openapi: 3.0.2
info:
  title: Paymentic APIs - Ledger
  version: 0.0.1
  description: Double Entry Ledger APIs
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
paths:
  /shelves:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestShelfCreation'
        required: true
      tags:
        - shelf
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/shelfCreated'
          description: Book created
        '401':
          $ref: '#/components/responses/res-401'
        '403':
          $ref: '#/components/responses/res-403'
        '422':
          $ref: '#/components/responses/res-422'
        '500':
          $ref: '#/components/responses/res-500'
      operationId: create-shelf
      summary: It creates a new shelf
  '/shelves/{id}':
    summary: It retrieves the book
    get:
      tags:
        - shelf
      responses:
        '200':
          $ref: '#/components/responses/res-200-shelf'
        '401':
          $ref: '#/components/responses/res-401'
        '403':
          $ref: '#/components/responses/res-403'
        '404':
          $ref: '#/components/responses/res-404'
        '500':
          $ref: '#/components/responses/res-500'
      operationId: get-shelf
      summary: It retrieves a book
    parameters:
      -
        name: id
        description: Book ID
        schema:
          type: string
        in: path
        required: true
  '/shelves/{id}/books':
    get:
      tags:
        - book
      parameters:
        -
          name: id
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          $ref: '#/components/responses/res-200-books'
        '401':
          $ref: '#/components/responses/res-401'
        '403':
          $ref: '#/components/responses/res-403'
        '500':
          $ref: '#/components/responses/res-500'
      operationId: get-books
  '/shelves/{id}/books/{bookId}/records':
    get:
      tags:
        - book
      parameters:
        -
          name: id
          schema:
            type: string
          in: path
          required: true
        -
          name: bookId
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          $ref: '#/components/responses/res-200-records'
        '401':
          $ref: '#/components/responses/res-401'
        '403':
          $ref: '#/components/responses/res-403'
        '500':
          $ref: '#/components/responses/res-500'
      operationId: get-records
components:
  schemas:
    error-401:
      title: error-401
      description: ''
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      example:
        code: PAY-001
        description: invalid authentication
    error-403:
      title: error-403
      description: ''
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      example:
        code: PAY-003
        description: insufficient permissions
    error-404:
      title: error-404
      description: ''
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      example:
        code: PAY-404
        description: resource not found
    error-422:
      title: error-422
      description: ''
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      example:
        code: PAY-002
        description: insufficient funds
    error-500:
      title: error-500
      description: ''
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      example:
        code: PAY-500
        description: internal server error
    shelf:
      title: Root Type for book
      description: ''
      type: object
      properties:
        id:
          type: string
        owner:
          type: object
          properties:
            id:
              type: string
        version:
          format: int32
          type: integer
      example:
        id: 5f21f211-c2f5-496e-b08b-2e8723b24483
        owner:
          id: 99eabea1-8dbd-437f-a706-a73162d5ba7b
        version: 10
    shelfCreated:
      title: Root Type for bookCreated
      description: ''
      type: object
      properties:
        id:
          type: string
      example:
        id: bfcf3d99-ec2d-4e86-9b1e-1b43c1cc7188
    requestShelfCreation:
      title: Root Type for requestBookCreation
      description: ''
      type: object
      properties:
        owner:
          type: object
          properties:
            id:
              type: string
      example:
        owner:
          id: 99eabea1-8dbd-437f-a706-a73162d5ba7b
    book:
      title: Root Type for book
      description: ''
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        version:
          format: int32
          type: integer
      example:
        id: f4b4f28f-56c4-4427-b772-0ef38f5653f8
        type: payments
        version: 10
    books:
      title: Root Type for books
      description: ''
      type: array
      items:
        type: object
        properties:
          id:
            type: string
          type:
            type: string
          version:
            format: int32
            type: integer
      example:
        -
          id: f4b4f28f-56c4-4427-b772-0ef38f5653f8
          type: payments
          version: 10
        -
          id: f3b4f28f-56c4-4427-b772-0ef38f5653f8
          type: pending
          version: 10
    record:
      title: Root Type for record
      description: ''
      type: object
      properties:
        id:
          type: string
        at:
          format: date
          type: string
        amount:
          format: double
          type: number
        currency:
          type: string
        type:
          type: string
      example:
        id: 944fd016-d9fc-4ff6-b69b-a80878a7a0d7
        at: '2023-02-21'
        amount: 112.21
        currency: BRL
        type: CREDIT
    records:
      title: Root Type for records
      description: ''
      type: array
      items:
        type: object
        properties:
          id:
            type: string
          at:
            format: date
            type: string
          amount:
            format: double
            type: number
          currency:
            type: string
          type:
            type: string
      example:
        -
          id: 944fd016-d9fc-4ff6-b69b-a80878a7a0d7
          at: '2023-02-21'
          amount: 112.21
          currency: BRL
          type: CREDIT
        -
          id: 944fd016-d9fc-4ff6-b69b-a80878a7a0d7
          at: '2023-02-21'
          amount: 112.21
          currency: BRL
          type: DEBIT
  responses:
    res-500:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error-500'
      description: ''
    res-401:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error-401'
      description: ''
    res-403:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error-403'
      description: ''
    res-422:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error-422'
      description: ''
    res-400:
      content:
        application/json: {}
      description: ''
    res-404:
      content:
        application/json: {}
      description: ''
    res-200-shelf:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/shelf'
      description: ''
    res-200-books:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/books'
      description: ''
    res-200-records:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/records'
      description: ''
tags:
  -
    name: shelf
    description: Shelf API
  -
    name: book
    description: Book API